requirements.txt
# Core Data Science Libraries
pandas
numpy
networkx

# Machine Learning & NLP
sentence-transformers
scikit-learn

# API Integration
groq
google-generativeai

# Web Framework
fastapi
uvicorn[standard]
pydantic

# Utilities
python-dotenv
